Ajustement de certains type, ajout de commentaires

+++ Compte.class.php
@@ -1,18 +1,19 @@
 <?php  
     class Compte{
         private string $_libellé;
-        private int $_solde;
+        private float $_solde;
         private string $_deviseMonétaire;
         private Titulaire $_titulaire;
 
-        public  function __construct(string $libellé, int $solde, string $deviseMonétaire, Titulaire $titulaire){
+        /* Fonction __construct de la classe */
+        public  function __construct(string $libellé, float $solde, string $deviseMonétaire, Titulaire $titulaire){
             $this->_libellé = $libellé;
             $this->_solde = $solde;
             $this->_deviseMonétaire = $deviseMonétaire;
             $this->_titulaire = $titulaire;
             $this->_titulaire->setComptesBancaires($this);
         }
-
+        /* Getter et Setter du libellé */
         public function getLibellé() : string{
             return $this->_libellé;
         }
@@ -20,6 +21,7 @@
             $this->_libellé = $libellé;
         }
 
+        /* Getter et Setter du solde */
         public function getSolde() : string{
             return $this->_solde;
         }
@@ -27,6 +29,7 @@
             $this->_solde = $solde;
         }
 
+        /* Getter et Setter de la devise monétaire */
         public function getDeviseMonétaire() : string{
             return $this->_deviseMonétaire;
         }
@@ -34,6 +37,7 @@
             $this->_deviseMonétaire = $deviseMonétaire;
         }
 
+        /* Getter et Setter du titulaire */
         public function getTitulaire() : Titulaire{
             return  $this->_titulaire;
         }
@@ -41,7 +45,8 @@
             $this->_titulaire = $titulaire;
         }
 
-        public function créditer(int $somme){
+        /* Méthode qui sert à créditer un compte */
+        public function créditer(float $somme){
             $this->_solde += $somme;
             /*
             echo "Vous avez créditer le compte " . $this->getLibellé() . " de " . $this . " de $somme " . $this->getDeviseMonétaire() . ", il contient maintenant : " . $this->getSolde() . " " . $this->getDeviseMonétaire() . "<br>";
@@ -51,7 +56,8 @@
             */
         }
 
-        public function débiter(int $somme){
+        /* Méthode qui sert à débiter un compte */
+        public function débiter(float $somme){
             $this->_solde -= $somme;
             /*
             echo "Vous avez débiter ce compte " . $this->_libellé . " de " . $this . " de $somme " . $this->getDeviseMonétaire() . ", il contient maintenant : " . $this->getSolde() . " " . $this->getDeviseMonétaire() . "<br>";
@@ -61,7 +67,8 @@
             */
         }
 
-        public function virement(Compte $versCompte, int $somme){
+        /* Méthode qui sert à faire un virement de ce compte vers un autre */
+        public function virement(Compte $versCompte, float $somme){
             if($this->getSolde() < $somme){
                 echo "Impossible, la somme du virement est supérieur au solde du compte \"" . $this->getLibellé() . "\" appartenant à " . $this->getTitulaire() . ".<br>";
             }
@@ -70,10 +77,9 @@
                 $versCompte->créditer($somme);
                 echo "Vous avez effectuer un virement de $somme depuis le compte \"" . $this->getLibellé() . "\" appartenant à " . $this->getTitulaire() . " vers le compte \"" . $versCompte->getLibellé() . "\" appartenant à " . $versCompte->getTitulaire() . ".<br>";
             }
-            
-
         }
 
+        /* Méthode qui sert à retourner les détails du compte */
         public function détailCompte(){
             return "Voici les informations de ce compte bancaire : <br>
                         - Libellé : $this->_libellé<br>
@@ -81,6 +87,7 @@
                         - Titulaire : " . $this->getTitulaire() . "<br>";
         }
 
+        /* Méthode __toString de la classe */
         public function __toString(){
             return $this->_libellé;
         }
